fix tree Locations
